// src/app/modules/company/models/comany.ts
export type Company = {

    id?: string
    code?: string
    name?: string
    address?: string
    cityCode?: string
    cityName?: string
    wilayaName?: string //willaya code missed
    wilayaCode?: string //willaya code missed
    zipCode?: string
    email?: string
    number?: string
   



}